+++ src/function.js
@@ -1,6 +1,6 @@
 const register = [];
 
-export const correctRegiste = (name, lastname, birthday, email, password, confirmPassword) => {
+export const correctRegister = (name, lastname, birthday, email, password, confirmPassword) => {
   if (name === '' || lastname === '' || birthday === '' || email === '' || password === '' || confirmPassword === '') {
     console.log('funciona funcion');
     alert('Por favor completar campo NOMBRE');
