--- src/index.js
@@ -1,3 +1,0 @@
-import { example } from './example.js';
-
-example();
