+++ src/view-controller/router.js
@@ -1,9 +1,10 @@
-import { components } from '../view/index.js';
+import { components } from '../view-screen/index.js';
 
 export const changeTmp = (hash) => {
   const id = hash.split('/')[1];
   const sectionMain = document.getElementById('container');
   sectionMain.innerHTML = '';
+
 
   switch (hash) {
     case '':
@@ -13,7 +14,7 @@ export const changeTmp = (hash) => {
     case '#/register':
     case '#/foro':
     case '#/contactenos':
-    { return sectionMain.appendChild(components[id]()); }
+    {return sectionMain.appendChild(components[id]()); }
     default:
     { return sectionMain.appendChild(components.different()); }
   }
