+++ src/view-screen/contacto.js
@@ -1,0 +1,8 @@
+export default () => {
+  const viewContacto = 'ejemplo';
+
+  const divElemt = document.createElement('div');
+  divElemt.classList.add('position');
+  divElemt.innerHTML = viewContacto;
+  return divElemt;
+};
