+++ src/view-screen/index.js
@@ -5,13 +5,10 @@ import Contactenos from './contacto.js';
 import Different from './404.js';
 
 
-const components = {
+export const components = {
   home: Home,
   register: Register,
   foro: Foro,
   contacto: Contactenos,
   different: Different,
 };
-
-
-export { components };
