--- src/view/foro.js
@@ -1,8 +1,0 @@
-export default () => {
-    const viewForo = `aqui va un div con la información`;
-
-    const divElemt = document.createElement('div');
-    divElemt.classList.add('position')
-    divElemt.innerHTML = viewForo;
-    return divElemt;
-};
